moved the fetch to the reducer, made a new home-page
