+++ code/src/App.js
@@ -4,8 +4,8 @@ import { ScanBarcode } from "components/ScanBarcode"
 import { Provider, useDispatch } from 'react-redux'
 import { configureStore, combineReducers } from '@reduxjs/toolkit'
 import { ui } from 'reducers/ui'
-import { products } from 'reducers/products'
-// import { onDetected } from 'reducers/products'
+import { products, fetchProductData } from 'reducers/products'
+import { Home } from './components/Home'
 
 const reducer = combineReducers({
   ui: ui.reducer,
@@ -13,7 +13,6 @@ const reducer = combineReducers({
 })
 
 export const store = configureStore({ reducer })
-const dispatch = useDispatch()
 
 // const onDetected = (code) => {
 //   console.log(`Code: ${code}`)
@@ -24,31 +23,40 @@ const dispatch = useDispatch()
 //     })
 // };
 
+
+
+
+
 export const App = () => {
+
+
+
   return (
     <div>
       <Provider store={store}>
-        {/*  <BrowserRouter>
-          <Switch>
-            <Route path='/' exact> */}
-        <label>
-          {" "}
-              Test codes here:{" "}
-          <input type="text" onChange={(e) => onDetected(e.target.value)}></input>
-        </label>
-        <p>
-          {" "}
-              Use the field above to test barcodes manually and keep an eye on your
-              console in the browser. i.e. Type 7311070347272 - Pågen Gifflar. Yum
-            </p>
-        <ScanBarcode />
-        {/* </Route>
-            <Route path='/product/:productId'>
-               <ScanBarcode />
-            </Route>
-          </Switch>
-        </BrowserRouter> */}
+        <Home />
       </Provider>
     </div >
   );
 };
+
+        // {/*  <BrowserRouter>
+        //   <Switch>
+        //     <Route path='/' exact> */}
+        //     <label>
+        //     {" "}
+        //         Test codes here:{" "}
+        //     <input type="text" onChange={(e) => onDetected(e.target.value)}></input>
+        //   </label>
+        //   <p>
+        //     {" "}
+        //         Use the field above to test barcodes manually and keep an eye on your
+        //         console in the browser. i.e. Type 7311070347272 - Pågen Gifflar. Yum
+        //       </p>
+        //   <ScanBarcode />
+        //   {/* </Route>
+        //       <Route path='/product/:productId'>
+        //          <ScanBarcode />
+        //       </Route>
+        //     </Switch>
+        //   </BrowserRouter> */}
