+++ code/src/components/ScanBarcode.js
@@ -1,8 +1,11 @@
 import React, { useState } from 'react'
 import { BarcodeScanner } from 'components/BarcodeScanner'
+import { useDispatch } from 'react-redux'
 
-export const ScanBarcode = () => {
+
+export const ScanBarcode = ({ onDetected }) => {
   const [showScanner, setShowScanner] = useState(false)
+  // const dispatch = useDispatch()
 
   return (
     <>
@@ -16,6 +19,7 @@ export const ScanBarcode = () => {
         <BarcodeScanner onDetected={(code) => {
           console.log('Got barcode', code)
           setShowScanner(false)
+          onDetected(code)
         }} />
       )}
     </>
